Final Commit (for now) + Application
Final commit for Finfillment to make sure the code is the upmost updated it could be. WARNING, this current code DOES NOT WORK. But, I did include a Unity Application that vaguely works so you can test the saving and loading function of the store buying. But a lot of the code should just be looked at on GitHub instead of downloading it.